Move `RealmPanel` to `ui` directory
