+++ src/shared/components/Modal/ClickableModalOverlay.tsx
