+++ src/ui/Island/RealmPanel/RealmDetailsPanel/RealmDetailsRewards.tsx
