+++ src/ui/Onboarding/ConnectWallet.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import OnboardingModal from "shared/components/Modals/OnboardingModal"
+import OnboardingModal from "ui/Island/Modals/OnboardingModal"
 import { useConnect } from "shared/hooks"
 
 export default function ConnectWallet() {
