+++ src/ui/Onboarding/Loader.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import GoldenCircleSpinner from "shared/components/Loaders/GoldenCircleSpinner"
 import LoadingText from "shared/components/Loaders/LoadingText"
-import OnboardingModal from "shared/components/Modals/OnboardingModal"
+import OnboardingModal from "ui/Island/Modals/OnboardingModal"
 
 export default function OnboardingModalLoader() {
   return (
